+++ lib/database/models/auth/auth_code_.dart
@@ -2,7 +2,7 @@ import "dart:math";
 
 import "package:mongo_dart/mongo_dart.dart";
 import "package:rpmtw_server/database/database.dart";
-import "package:rpmtw_server/database/base_models.dart";
+import 'package:rpmtw_server/database/db_model.dart';
 import "package:rpmtw_server/database/index_fields.dart";
 import "package:rpmtw_server/utilities/utility.dart";
