+++ lib/database/models/comment/comment.dart
@@ -1,5 +1,5 @@
 import "package:rpmtw_server/database/database.dart";
-import "package:rpmtw_server/database/base_models.dart";
+import 'package:rpmtw_server/database/db_model.dart';
 import "package:rpmtw_server/database/models/comment/comment_type.dart";
 import "package:rpmtw_server/database/index_fields.dart";
 import "package:rpmtw_server/database/model_field.dart";
