+++ lib/database/models/minecraft/minecraft_mod.dart
@@ -1,6 +1,6 @@
 import "package:rpmtw_server/database/database.dart";
 
-import "package:rpmtw_server/database/base_models.dart";
+import 'package:rpmtw_server/database/db_model.dart';
 import "package:rpmtw_server/database/index_fields.dart";
 import "package:rpmtw_server/database/models/minecraft/relation_mod.dart";
 import "package:rpmtw_server/database/models/minecraft/minecraft_version.dart";
