+++ lib/database/models/minecraft/minecraft_version_manifest.dart
@@ -4,7 +4,7 @@ import "package:http/http.dart";
 import "package:mongo_dart/mongo_dart.dart";
 
 import "package:rpmtw_server/database/database.dart";
-import "package:rpmtw_server/database/base_models.dart";
+import 'package:rpmtw_server/database/db_model.dart';
 import "package:rpmtw_server/database/index_fields.dart";
 import "package:rpmtw_server/database/models/minecraft/minecraft_version.dart";
 import "package:rpmtw_server/utilities/utility.dart";
