+++ lib/database/models/minecraft/rpmwiki/wiki_change_log.dart
@@ -2,7 +2,7 @@ import "dart:convert";
 
 import "package:mongo_dart/mongo_dart.dart";
 import "package:rpmtw_server/database/database.dart";
-import "package:rpmtw_server/database/base_models.dart";
+import 'package:rpmtw_server/database/db_model.dart';
 import "package:rpmtw_server/database/index_fields.dart";
 import "package:rpmtw_server/database/models/minecraft/minecraft_mod.dart";
