+++ lib/database/models/translate/source_text.dart
@@ -4,7 +4,7 @@ import "package:intl/locale.dart";
 import "package:collection/collection.dart";
 
 import "package:rpmtw_server/database/database.dart";
-import "package:rpmtw_server/database/base_models.dart";
+import 'package:rpmtw_server/database/db_model.dart';
 import "package:rpmtw_server/database/index_fields.dart";
 import "package:rpmtw_server/database/models/minecraft/minecraft_version.dart";
 import "package:rpmtw_server/database/model_field.dart";
