+++ lib/database/models/universe_chat/universe_chat_message.dart
@@ -1,7 +1,7 @@
 import "dart:io";
 
 import "package:rpmtw_server/database/database.dart";
-import "package:rpmtw_server/database/base_models.dart";
+import 'package:rpmtw_server/database/db_model.dart';
 import "package:rpmtw_server/database/index_fields.dart";
 
 class UniverseChatMessage extends DBModel {
